+++ packages/vtable-plugins/src/column-series.ts
@@ -7,22 +7,23 @@ export interface ColumnSeriesOptions {
   generateColumnTitle?: (index: number) => string;
   generateColumnField?: (index: number) => string;
   /**
-   * 是否自动扩展列
+   * 是否自动扩展列, 配置监听键值（键盘事件的key）
    * @default true
    */
-  autoExtendColumn?: boolean;
+  autoExtendColumnTriggerKeys?: ('ArrowRight' | 'Tab')[];
 }
 /**
  * 生成列序号标题的插件
  */
 export class ColumnSeriesPlugin implements VTable.plugins.IVTablePlugin {
-  id = 'column-series';
+  id = `column-series-${Date.now()}`;
+  name = 'Column Series';
   runTime = [VTable.TABLE_EVENT_TYPE.BEFORE_INIT, VTable.TABLE_EVENT_TYPE.BEFORE_KEYDOWN];
   pluginOptions: ColumnSeriesOptions;
   table: VTable.ListTable;
   columns: { field?: string; title: string }[] = [];
   constructor(pluginOptions: ColumnSeriesOptions) {
-    this.pluginOptions = Object.assign({ columnCount: 100, autoExtendColumn: true }, pluginOptions);
+    this.pluginOptions = Object.assign({ columnCount: 100 }, pluginOptions);
   }
   run(...args: any[]) {
     if (args[1] === VTable.TABLE_EVENT_TYPE.BEFORE_INIT) {
@@ -36,11 +37,8 @@ export class ColumnSeriesPlugin implements VTable.plugins.IVTablePlugin {
     } else if (args[1] === VTable.TABLE_EVENT_TYPE.BEFORE_KEYDOWN) {
       const eventArgs = args[0];
       const e = eventArgs.event;
-      if (e.key === 'ArrowRight') {
-        if (
-          this.pluginOptions.autoExtendColumn &&
-          this.table.stateManager.select.cellPos.col === this.table.colCount - 1
-        ) {
+      if (this.pluginOptions.autoExtendColumnTriggerKeys?.includes(e.key)) {
+        if (this.table.stateManager.select.cellPos.col === this.table.colCount - 1) {
           this.table.addColumn(this.generateColumn(this.table.colCount - 1) as VTable.ColumnDefine);
         }
       }