+++ packages/vtable-plugins/src/excel-edit-cell-keyboard.ts
@@ -1,5 +1,3 @@
-import type { CellRange } from '@visactor/vtable/es/ts-types';
-import type { BaseTableAPI } from '@visactor/vtable/es/ts-types/base-table';
 import * as VTable from '@visactor/vtable';
 import type { TableEvents } from '@visactor/vtable/src/core/TABLE_EVENT_TYPE';
 import type { EventArg } from './types';
@@ -10,7 +8,8 @@ export type IExcelEditCellKeyboardPluginOptions = {
 };
 
 export class ExcelEditCellKeyboardPlugin implements VTable.plugins.IVTablePlugin {
-  id = 'excel-edit-cell-keyboard';
+  id = `excel-edit-cell-keyboard-${Date.now()}`;
+  name = 'Excel Edit Cell Keyboard';
   runTime = [VTable.TABLE_EVENT_TYPE.INITIALIZED];
   table: VTable.ListTable;
   pluginOptions: IExcelEditCellKeyboardPluginOptions;